+++ packages/websocket/lib/websocket.js
@@ -1,0 +1,5 @@
+"use strict";
+
+// @skyjs/websocket entry: re-exports the RFC 6455 implementation.
+
+module.exports = require("./websocket-core.js");